requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
jinja2==3.1.6
pydantic==2.12.5
python-dotenv==1.2.1

supabase==2.27.2
httpx==0.28.1
httpcore==1.0.9
pyjwt==2.10.1
python-jose[cryptography]==3.5.0
pydantic-settings==2.12.0